feat(admin-ui): Avoid duplication initAudit generator function in sagas
