+++ plugins/auth-server/redux/sagas/JsonConfigSaga.js
@@ -16,6 +16,7 @@ import { PATCH, FETCH } from '../../../../app/audit/UserActionType'
 import { postUserAction } from '../../../../app/redux/api/backend-api'
 
 const JansConfigApi = require('jans_config_api')
+import { initAudit } from '../../plugin-selector'
 
 function* newFunction() {
   const token = yield select((state) => state.authReducer.token.access_token)
@@ -24,19 +25,6 @@ function* newFunction() {
     getClient(JansConfigApi, token, issuer),
   )
   return new JsonConfigApi(api)
-}
-
-function* initAudit() {
-  const auditlog = {}
-  const client_id = yield select((state) => state.authReducer.config.clientId)
-  const ip_address = yield select((state) => state.authReducer.location.IPv4)
-  const userinfo = yield select((state) => state.authReducer.userinfo)
-  const author = userinfo ? userinfo.family_name || userinfo.name : '-'
-  auditlog['client_id'] = client_id
-  auditlog['ip_address'] = ip_address
-  auditlog['author'] = author
-  auditlog['status'] = 'succeed'
-  return auditlog
 }
 
 export function* getJsonConfig() {
