+++ plugins/auth-server/redux/sagas/OAuthScopeSaga.js
@@ -40,6 +40,7 @@ import {
 import { postUserAction } from '../../../../app/redux/api/backend-api'
 
 const JansConfigApi = require('jans_config_api')
+import { initAudit } from '../../plugin-selector'
 
 function* newFunction() {
   const token = yield select((state) => state.authReducer.token.access_token)
@@ -48,18 +49,6 @@ function* newFunction() {
     getClient(JansConfigApi, token, issuer),
   )
   return new ScopeApi(api)
-}
-function* initAudit() {
-  const auditlog = {}
-  const client_id = yield select((state) => state.authReducer.config.clientId)
-  const ip_address = yield select((state) => state.authReducer.location.IPv4)
-  const userinfo = yield select((state) => state.authReducer.userinfo)
-  const author = userinfo ? userinfo.family_name || userinfo.name : '-'
-  auditlog['client_id'] = client_id
-  auditlog['ip_address'] = ip_address
-  auditlog['author'] = author
-  auditlog['status'] = 'succeed'
-  return auditlog
 }
 
 export function* getScopeByInum() {
